# product-service/src/main/java/com/kika/product_service/dto/ProductRequestDto.java
package com.kika.product_service.dto;

import lombok.*;

import java.math.BigDecimal;
import java.util.List;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class ProductRequestDto {
    private String name;
    private String description;
    private BigDecimal price;
    private String category;
    private boolean published;
    private List<String> tags;
    private List<String> images;
}